requirements.txt
pypdf==3.17.2
beautifulsoup4
requests
pillow==10.2.0
